Check environ before uninstalling
The following moves the environ checking further up the stack, so
that we exit out earlier before continuing onwards.
